// react-redux/counter-app/src/stories/sidebar/interface.ts
export interface DetailsData {
  id: string;
  partsNew: string;
  partsAM: string;
  partsUsed: string;
  glass: string;
  partsOthers: string;
  paint: string;
  body: string;
  mech: string;
  frame: string;
  misc: string;
  detail: string;
  labourOther: string;
  sublet: string;
  customLabour: string;
  customTax: string;
  payableTaxes: string;
}